refactor: Reorganize application layer with improved separation of concerns
- Split monolithic main.go (442 lines) into focused files:
  - main.go: Clean application entry point and orchestration (68 lines)
  - config.go: Configuration parsing and environment handling (126 lines)
  - server.go: HTTP server setup and middleware configuration (92 lines)
  - infrastructure.go: NATS, email, and JWT infrastructure setup (215 lines)

- Improved code organization:
  - Separated configuration concerns into dedicated config.go
  - Isolated HTTP server setup logic in server.go
  - Consolidated infrastructure setup in infrastructure.go
  - Maintained clean separation of concerns and single responsibility

- Enhanced configuration management:
  - Extracted email configuration into focused parseEmailConfig function
  - Organized email config into structured emailConfig type
  - Improved environment variable parsing with clear defaults

- Fixed logging consistency:
  - Removed duplicate errKey constant definition
  - Updated all error logging to use logging.ErrKey from logging package
  - Eliminated redundant constant declarations across application layer

- No functional changes - maintains backward compatibility
- Reduced main.go complexity by 85% (442 → 68 lines)
- Improved maintainability and readability
